requirements.txt
Pillow==10.0.0
requests
